refactor: split prune.ts into inject.ts and prune.ts for SRP
